requirements.txt
# Core dependencies
playwright>=1.32.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
